+++ blocks/src/membership/block.json
@@ -1,0 +1,31 @@
+{
+	"$schema": "https://schemas.wp.org/trunk/block.json",
+	"apiVersion": 3,
+	"name": "pmpro/membership",
+	"title": "Membership Required Block",
+	"category": "pmpro",
+	"icon": "visibility",
+	"description": "Nest blocks within this wrapper to control the inner block visibility by membership level or for non-members only.",
+	"attributes" : {
+		"levels": {
+			"type": "array",
+			"default":[]
+		},
+		"uid": {
+			"type": "string",
+			"default":""
+		},
+		"show_noaccess": {
+			"type": "string",
+			"default": "0"
+		}
+	},
+	"supports": {
+		"html": false
+	},
+	"textdomain": "paid-memberships-pro",
+	"editorScript": "file:./index.js",
+	"editorStyle": "file:./index.css",
+	"style": "file:./style-index.css",
+	"render": "file:./render.php"
+}
